requirements.txt
flask
flask-sqlalchemy
flask-httpauth